Ninety-seventh Amendment
See the PDF: http://legislative.gov.in/sites/default/files/amend97.pdf
SOR: http://legislative.gov.in/sites/default/files/99%20SOR.pdf

SOR Follows:

STATEMENT OF OBJECTS AND REASONS
The co-operative sector, over the years, has made significant contribution to various
sectors of national economy and has achieved voluminous growth. However, it has shown
weaknesses in safeguarding the interests of the members and fil.!fiIment of objects for which
these institutions were organised. There have been instances where elections have been
postponed indefinitely and nominated office bearers or administrators remaining in-charge
of these institutions for a long time. This reduces the accountability of the management of
co-operative societies to their members. Inadequate professionalism in management in many
of the co-operative institutions has led to poor services and low productivity. Co-operatives
need to run on well established democratic principles and elections held on time and in a free
and fair manner. Therefore, there is a need to initiate fundamental reforms to revitalize these
institutions in order to ensure their contribution in the economic development of the country
and to serve the interests of members and public at large and also to ensure their autonomy,
democratic functioning and professional management.
2. The "co-operative societies" is a subject enumerated in Entry 32 of the State List of
the Seventh Schedule of the Constitution and the State Legislatures have accordingly enacted
legislations on co-operative societies. Within the framework of State Acts, growth of
co-operatives on large scale was envisaged as part of the efforts for securing social and
economic justice and equitable distribution of the fruits of development. It has, however,
been experienced that in spite of considerable expansion of co-operatives, their performance
in qualitative terms has not been up to the desired level. Considering the need for reforms in
the Co-operative Societies Acts of the States, consultations with the State Governments
have been held at several occasions and in the conferences of State Co-operative Ministers.
A strong need has been felt for amending the Constitution so as to keep the co-operatives
free from unnecessary outside interferences
and also to ensure, their autonomous
organisational set up and their democratic functioning.
3. The Central Government is committed to ensure that the co-operative societies in
the country function in a democratic, professional, autonomous and economically sound
manner. With a view to bring the necessary reforms, itis proposed to incorporate a new Part
in the Constitution so- as to provide for certain provisions covering the vital. aspects of
working of co-operative societies like democratic, autonomous and professional functioning.
A new article is also proposed to be inserted in Part IV of the Constitution (Directive Principles
of State Policy) for the States to endeavour to promote voluntary formation, autonomous
functioning, democratic control and professional management of cooperative societies. The
proposed new Part in the Constitution, inter alia, seeks to empower the Parliament in respect
of multi-State co-operative societies and the State Legislatures in case of other co-operative
societies to make appropriate law, laying down the following matters, namely:(a) provisions for incorporation, regulation arid winding up of co-operative
societies based on the principles of democratic member-control, member-economic
participation and autonomous functioning;
(b) specifying the maximum number of directors of a co-operative society to be
not exceeding twenty-one members;
(c) providing for a fixed term offive years from the date of election in respect of
the elected members of the board and its office bearers;
(d) providing for a maximum time limit of six months during which a board of

directors of co-operative

society could be kept under supersession or suspension;

(e) providing for independent professional

audit;

(f) providing for right of information to the members of the co-operative societies;
6

7
(g) empowering the State Governments
and accounts of co-operative societies;

to obtain periodic

reports of activities

(h) providing for the reservation of one seat for the Scheduled Castes or the
Scheduled Tribes and two seats for women on the board of every co-operative society,
which have individuals as members from such categories;

(i) providing for offences relating to co-operative societies and penalties in
respect of such offences.

4. It is expected that these provisions will not only ensure the autonomous and
democratic functioning of co-operatives, but also ensure the accountability of management
to the members and other stakeholders and shall provide for deterrence for violation of the
provisions of the law.

5. The Bill seeks to achieve the above objectives.

NEW DELHI:

The 11thNovember, 2009.
SHARAD PAWAR