Ninety-ninth amendment (#1)
* Ninety-ninth amendment

Amendment: http://legislative.gov.in/sites/default/files/99th.pdf
SOR: http://legislative.gov.in/sites/default/files/99%20SOR.pdf

SOR Text follows:

STATEMENT OF OBJECTS AND REASONS
The Judges of the Supreme Court are appointed under clause (2) of article 124 and the
Judges of the High Courts are appointed under clause (1) of article 217 of the Constitution,
by the President. The Ad-hoc Judges and retired Judges for the Supreme Court are appointed
under clause (1) of article 127 and article 128 of the Constitution respectively. The appointment
of Additional Judges and Acting Judges for the High Court is made under article 224 and the
appointment of retired Judges for sittings of the High Courts is made under article 224A of
the Constitution. The transfer of Judges from one High Court to another High Court is made
by the President after consultation with the Chief Justice of India under clause (1) of article
222 of the Constitution.
2. The Supreme Court in the matter of the Supreme Court Advocates-on-Record
Association Vs. Union of India in the year 1993, and in its Advisory Opinion in the year 1998
in the Third Judges case, had interpreted clause (2) of article 124 and clause (1) of article 217
of the Constitution with respect to the meaning of “consultation” as “concurrence”.
Consequently, a Memorandum of Procedure for appointment of Judges to the Supreme
Court and High Courts was formulated, and is being followed for appointment.
3. After review of the relevant constitutional provisions, the pronouncements of the
Supreme Court and consultations with eminent Jurists, it is felt that a broad based National
Judicial Appointments Commission should be established for making recommendations for
appointment of Judges of the Supreme Court and High Courts. The said Commission would
provide a meaningful role to the judiciary, the executive and eminent persons to present their
view points and make the participants accountable, while also introducing transparency in
the selection process.
4. The Constitution (One Hundred and Twenty-first Amendment) Bill, 2014 is an enabling
constitutional amendment for amending relevant provisions of the Constitution and for
setting up a National Judicial Appointments Commission. The proposed Bill seeks to insert
new articles 124A, 124B and 124C after article 124 of the Constitution. The said Bill also
provides for the composition and the functions of the proposed National Judicial
Appointments Commission. Further, it provides that Parliament may, by law, regulate the
procedure for appointment of Judges and empower the National Judicial Appointments
Commission to lay down procedure by regulation for the discharge of its functions, manner
of selection of persons for appointment and such other matters as may be considered
necessary.
5. The proposed Bill seeks to broad base the method of appointment of Judges in the
Supreme Court and High Courts, enables participation of judiciary, executive and eminent
persons and ensures greater transparency, accountability and objectivity in the appointment
of the Judges in the Supreme Court and High Courts.
6. The Bill seeks to achieve the above objectives.
NEW DELHI;

RAVI SHANKAR PRASAD.
The 8th August, 2014.

* Add the resignation clause back