One Hundredth Amendment
Date on which the Act came into force: 31 July 2015

STATEMENT OF OBJECTS AND REASONS

India and Bangladesh have a common land boundary of approximately 4096.7 kms.
The India-East Pakistan land boundary was determined as per the Radcliffe Award of 1947.
Disputes arose out of some provisions in the Radcliffe award, which were sought to be
resolved through the Bagge Award of 1950. Another effort was made to settle these disputes
by the Nehru-Noon Agreement of 1958. However, the issue relating to division of Berubari
Union was challenged before the Hon'ble Supreme Court. To comply with the opinion rendered
by the Hon'ble Supreme Court of India, the Constitution (Ninth Amendment) Act, 1960 was
passed by the Parliament. Due to the continuous litigation and other political developments
at that time, the Constitution (Ninth Amendment) Act, 1960 could not be notified in respect
of territories in former East Pakistan (presently Bangladesh).

2. On 16th May, 1974, the Agreement between India and Bangladesh concerning the
demarcation of the land boundary and related matters was signed between both the
countries to find a solution to the complex nature of the border demarcation involved. This
Agreement was not ratified as it involved, inter alia, transfer of territory which requires a
Constitutional Amendment. In this connection, it was also required to identify the precise
area on the ground which would be transferred. Subsequently, the issues relating to
demarcation of un-demarcated boundary; the territories in adverse possession; and exchange
of enclaves were identified and resolved by signing a Protocol on 6th September, 2011,
which forms an integral part of the Land Boundary Agreement between India and Bangladesh,
1974. The Protocol was prepared with support and concurrence of the concerned State
Governments of Assam, Meghalaya, Tripura and West Bengal.

3. Accordingly, the Constitution (One Hundred and Ninteenth Amendment) Bill, 2013
which proposes to amend the First Schedule of the Constitution, for the purpose of giving
effect to the acquiring of territories by India and transfer of territories to Bangladesh through
retaining of adverse possession and exchange of enclaves, in pursuance of the aforesaid
Agreement of 1974 and its Protocol entered between the Governments of India and
Bangladesh.

4. The Bill seeks to achieve the above objects.

NEW DELHI;
The 26th February, 2013.

SOR URL: http://legislative.gov.in/sites/default/files/Constitution%20%28119th%29%20Bill%2C%202013.%20SOR.pdf
Amendment URL: http://egazette.nic.in/WriteReadData/2015/164169.pdf
Tables Manually generated using https://www.tablesgenerator.com/text_tables#